moved to new sniff
