+++ src/Sniffs/Attributes/SortedApiResourceOperationsSniff.php
@@ -16,13 +16,12 @@ use PhpParser\Node\Identifier;
 use PhpParser\ParserFactory;
 
 /**
- * The sniff checks if the keys in 'collectionOperations', 'itemOperations' and 'operations' attributes
- * within the 'ApiResource' annotation are properly sorted.
+ * The sniff checks 'operations' attributes within the 'ApiResource' annotation are properly sorted.
  *
- * The sorting criteria are defined in getRanks and getOperationsRanks methods.
+ * The sorting criteria are defined in the getRanks method.
  * If the keys are not sorted properly, the sniff will provide an option to fix the order.
  */
-final class SortedApiResourceOperationKeysSniff implements Sniff
+final class SortedApiResourceOperationsSniff implements Sniff
 {
     /**
      * @var string
@@ -32,7 +31,7 @@ final class SortedApiResourceOperationKeysSniff implements Sniff
     /**
      * @var array<string>
      */
-    private const API_RESOURCE_OPERATIONS_TO_PROCESS = ['collectionOperations', 'itemOperations', 'operations'];
+    private const API_RESOURCE_OPERATIONS_TO_PROCESS = ['operations'];
 
     /**
      * @var string
@@ -95,23 +94,6 @@ final class SortedApiResourceOperationKeysSniff implements Sniff
                 $items[$index] = $arrayItem;
             }
 
-            if ($arrayItem->value instanceof MethodCall) {
-                /** @var \PhpParser\Node\Expr\MethodCall $value */
-                $value = $arrayItem->value;
-                foreach ($value->args as $argIndex => $argument) {
-                    if ($argument instanceof Arg && $argument->value instanceof Array_) {
-                        /** @var \PhpParser\Node\Expr\ArrayItem[] $subItems */
-                        $subItems = $argument->value->items;
-                        /** @var int $startLine */
-                        $startLine = $argument->value->getAttribute('startLine');
-                        $argument->value->items = $this->fixMultiLineOutput($subItems, $startLine);
-                        $value->args[$argIndex] = $argument;
-                    }
-                }
-
-                $items[$index] = $arrayItem;
-            }
-
             if ($arrayItem->value instanceof New_) {
                 /** @var \PhpParser\Node\Expr\New_ $value */
                 $value = $arrayItem->value;
@@ -161,7 +143,7 @@ final class SortedApiResourceOperationKeysSniff implements Sniff
         return \strtolower(\trim($nodeKeyName, " \t\n\r\0\x0B\"'"));
     }
 
-    private function getOperationsRanks(ArrayItem $arrayItem): array
+    private function getRanks(ArrayItem $arrayItem): array
     {
         /** @var \PhpParser\Node\Expr\New_ $value */
         $value = $arrayItem->value;
@@ -188,42 +170,15 @@ final class SortedApiResourceOperationKeysSniff implements Sniff
     }
 
     /**
-     * @return array<bool|string>
-     */
-    private function getRanks(string $name): array
-    {
-        return [
-            \str_starts_with($name, 'export') === false,
-            \str_starts_with($name, 'get') === false,
-            \str_starts_with($name, 'post') === false,
-            \str_starts_with($name, 'put') === false,
-            \str_starts_with($name, 'patch') === false,
-            \str_starts_with($name, 'delete') === false,
-            \str_starts_with($name, 'activate') === false,
-            \str_starts_with($name, 'deactivate') === false,
-            $name,
-        ];
-    }
-
-    /**
      * @param \PhpParser\Node\Expr\ArrayItem[] $items
      *
      * @return \PhpParser\Node\Expr\ArrayItem[]
      */
     private function getSortedItems(array $items): array
     {
-        if ($this->isNotAssociativeOnly($items) === false) {
-            \uasort($items, function (ArrayItem $firstItem, ArrayItem $secondItem): int {
-                $firstName = $this->getArrayKeyAsString($firstItem);
-                $secondName = $this->getArrayKeyAsString($secondItem);
-
-                return $this->getRanks($firstName ?? '') <=> $this->getRanks($secondName ?? '');
-            });
-        }
-
         if ($this->isNotAssociativeOperationsArrayOnly($items)) {
             \uasort($items, fn (ArrayItem $firstItem, ArrayItem $secondItem): int =>
-                $this->getOperationsRanks($secondItem) <=> $this->getOperationsRanks($firstItem));
+                $this->getRanks($secondItem) <=> $this->getRanks($firstItem));
         }
 
         return $items;
@@ -241,20 +196,6 @@ final class SortedApiResourceOperationKeysSniff implements Sniff
         }
 
         return false;
-    }
-
-    /**
-     * @param \PhpParser\Node\Expr\ArrayItem[] $items
-     */
-    private function isNotAssociativeOnly(array $items): bool
-    {
-        $isNotAssociative = 1;
-
-        foreach ($items as $arrayItem) {
-            $isNotAssociative &= $arrayItem->key === null;
-        }
-
-        return (bool)$isNotAssociative;
     }
 
     /**
