+++ tests/Sniffs/Attributes/SortedApiResourceOperationsSniff/Fixture/Correct/correct.php.inc
